requirements.txt
discord.py
pyserial
